+++ async_upnp_client/advertisement.py
@@ -17,6 +17,7 @@ from async_upnp_client.ssdp import (
 )
 
 _LOGGER = logging.getLogger(__name__)
+_LOGGER_TRAFFIC_SSDP = logging.getLogger("async_upnp_client.traffic.ssdp")
 
 
 class UpnpAdvertisementListener:
@@ -68,9 +69,12 @@ class UpnpAdvertisementListener:
         self, request_line: str, headers: MutableMapping[str, str]
     ) -> None:
         """Handle data."""
-        _LOGGER.debug(
+        _LOGGER_TRAFFIC_SSDP.debug(
             "UpnpAdvertisementListener._on_data: %s, %s", request_line, headers
         )
+        if headers.get("MAN") == '"ssdp:discover"':
+            # Ignore discover packets.
+            return
         if "NTS" not in headers:
             _LOGGER.debug("Got unknown packet: %s, %s", request_line, headers)
             return
@@ -86,9 +90,11 @@ class UpnpAdvertisementListener:
 
     async def async_start(self) -> None:
         """Start listening for notifications."""
+        _LOGGER.debug("Start listening for notifications")
         self._transport, _ = await self._connect
 
     async def async_stop(self) -> None:
         """Stop listening for notifications."""
+        _LOGGER.debug("Stop listening for notifications")
         if self._transport:
             self._transport.close()