+++ async_upnp_client/search.py
@@ -18,6 +18,7 @@ from async_upnp_client.ssdp import (
 )
 
 _LOGGER = logging.getLogger(__name__)
+_LOGGER_TRAFFIC_SSDP = logging.getLogger("async_upnp_client.traffic.ssdp")
 
 
 async def async_search(
@@ -50,10 +51,15 @@ async def async_search(
     async def on_connect(transport: DatagramTransport) -> None:
         """Handle connection made."""
         packet = build_ssdp_search_packet(target_data, timeout, service_type)
+        _LOGGER.debug("Sending M-SEARCH packet")
+        _LOGGER_TRAFFIC_SSDP.debug("Sending M-SEARCH packet: %s", packet)
         transport.sendto(packet, target)
 
-    async def on_data(_: str, headers: MutableMapping[str, str]) -> None:
+    async def on_data(request_line: str, headers: MutableMapping[str, str]) -> None:
         """Handle data."""
+        _LOGGER.debug(
+            "Received response, request line: %s, headers: %s", request_line, headers
+        )
         headers["_source"] = "search"
         if target_ip:
             if headers["_address"].partition(":")[0] != f"{str(target_ip)}":